Add guard clause to unwrap optional parameter
The overridden method, `willMoveToSuperview`  in
AutoCompleteTextfield.swift is passed nil when used in a navigation
stack.

When the view with the autocomplete feature is in a child view of a
navigation stack, a user who taps the "back" button will encounter a
fatal error caused by force-unwrapping the optional parameter in
question.

Resolves #13

Use optional unwrapping